Update index.html
Tried my best to update the visualization logic of this app. It still needs some work, will commit soon. Will improve the visualization logic of the app. Will cross check if everything regarding all the things are correct in the next commit. And that should be more than enough.

We will have to work on the create logic by tonight and complete it.

Thank you.